+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/src/client_session_helper.py
