+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/src/main.py
