+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/src/requirements.txt
