+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/src/sts_helper.py
