+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/test/functional/ntest-no-qualifier.yaml
