+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/test/functional/ntest-unapproved-role.yaml
