+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/test/functional/ntest-wrong-qualifier.yaml
