+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/test/functional/ntest-xstack.yaml
