+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/test/functional/test-execution.yaml
