+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/test/unit/conftest.py
