+++ lambda/custom_resources/CTE_CrossAccountCloudFormation/test/unit/test_client_session_helper.py
