+++ lambda/custom_resources/CTE_InvokeCreateAccountFn/test/functional/cfn-sdlc-deploy-test.yaml
