+++ lambda/stepfunctions/CTE_GetAccountStatusFn/src/helper.py
