--- lambdas/custom_resources/CTE_CrossAccountCloudFormation/src/.python-version
@@ -1,1 +1,0 @@
-3.9.13
