--- lambdas/custom_resources/serverless.yaml
@@ -1,31 +1,0 @@
-# Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-Transform: "AWS::Serverless-2016-10-31"
-Description: Custom Resource to run a decoupled CloudFormation Stack (SAM)
-
-Resources:
-  rCTECrossAccountCloudFormationFn:
-    Type: AWS::Serverless::Function
-    Properties:
-      Handler: main.lambda_handler
-      Runtime: python3.9
-      FunctionName: CTE_CrossAccountCloudFormation
-      Description: This function will allow you to deploy a CloudFormation Stack across accounts.
-      Timeout: 900
-      CodeUri: CTE_CrossAccountCloudFormation/src
-      Policies:
-        - AdministratorAccess
-
-  rCTECrossAccountCloudFormationFnPermission:
-    Type: AWS::Lambda::Permission
-    Properties:
-      FunctionName: !GetAtt rCTECrossAccountCloudFormationFn.Arn
-      Action: lambda:InvokeFunction
-      Principal: cloudformation.amazonaws.com
-
-  rCTECrossAccountCloudFormationFnLogs:
-    Type: AWS::Logs::LogGroup
-    Properties:
-      LogGroupName: !Sub "/aws/lambda/${rCTECrossAccountCloudFormationFn}"
-      RetentionInDays: 7
